Clean up GPTService logic (#8)
We should be using GPTService exclusively at this point, so let's remove
all of the direct calls to openAI and other unused code.

Additionally, let's switch to gpt-service-2.boosted.ai. This is the
newer implementation of GPTService which hash response caching enabled.
This should help a lot with testing, since any query that we've already
seen (scoped to model_id, max_tokens and temperature) will not need to
hit OpenAI or Anthropic at all. This can be disabled if desired by
modifying the metadata in the client request to include a tuple with
they key "nocache".

Because of this caching logic, we can actually enable test_planner.