[PS24-1612] Refactor service level test (#67)
Currently, there is a "service level integration test", test_service.py.
This test has a lot of mocks and has been quite flaky. It submits jobs
to prefect which creates even more instability. That being said, the
logic in the test itself is solid.

This change aims to make the service level integration test much more
stable. Add a new test, TestAgentServiceImpl, that is designed to
replace the old test. Instead of hitting the full FastAPI application
directly with http requests, all of the handler logic has been moved out
of applciation.py and into the AgentServiceImpl class. This class will
own the external connections, like DB, prefect and GPTService. This
makes it much easier to setup a test that does NOT hit prefect, since we
can control the logic used to submit tasks without needing any usage of
the "mock" library at all. The structure of the test is based off of the
structure of the integrations tests that are currently present in
GRPC Service repos. Instead of inheriting from IsolatedAsyncioTestCase,
which causes problems with sharing connections across test cases, this
test manages its own event loop directly.

Additionally, since the auth logic is not present in AgentServiceImpl,
it is much easier to impersonate users for the  purposes of testing.

A side effect of this change is that we can now share a single
connection to GPTService across the app. Additionally, this will make it
much easier to switch to AsyncPostgresBase since we have more coverage.