[PS24-1667] Add event logging for tools (#98)
* [PS24-1667] Add event logging for tools

We need to know when a tool gets invoked, what arugments it was invoked
with, how long it took to run, who invoked it, if there were any errors
etc. This change adds event_logging with all of this info. Later, we
will create tables in Clickhouse to analyze the data.

* PR feedback