[WC-124] Add regression test suite (#159)
When a new tool is added, the prompt that gets sent to the LLM to
generate the execution plan for a user prompt will be modified to
include the new tool. Because of this, it is possible to break the
exeuction plan logic by adding a new tool. Currently it is very
difficult to know if adding a new tool will cause any issues. We'd like
to add a regression test that will alert us when code changes cause
breaking changes to the execution plan logic.

The existing way to test the addition of a tool is to use the
run_agent_offline.py script. This script takes a user prompt as input
and will generate the execution plan locally, then actually execute the
execution plan. This change adds a test,
regression_test/test_regression.py that is heavily inspired by the
offline script. However, this test contains a curated set of "user
prompts" that we'd like to track over time. Just like the offline
script, this test will generate the execution plan for the prompt and
then execute the execution plan. However, this test will also log events
to Clickhouse that will allow tracking changes over time.

Note that tests under regression_test/ will NOT be run on CI. These are
flaky, and we do not (yet) want to have assertions on the plans
generated since they are changing frequently. Instead, these tests will
get launched when we deploy new agent-service changes to DEV, or
manually if desired.