Set query_seconds to 0.5 in prefect_serve.py (#197)
There is overhead involved in creating tasks in prefect. The deployment
runner that we currently use with the "serve" method from the prefect
API will poll the prefect server to see if any tasks are avialable. The
default poll interval is 10 seconds. This means that it will check for
work every 10 seconds. This can cause delays. Let's decrease this
interval to 0.5 seconds - this should allow tasks to get picked up much
more quickly.