[WC-225] Prefect latency refactor (#210)
* [WC-225] Prefect latency refactor

We've been noticing a lot of latency when running tasks in prefect.
Specifically, when tasks are submitted via the run_deployment method
from the prefect API it can take 15+ seconds to actually start running
the task. It has been quite difficult to find a way to use
run_deployment and also reduce this latency.

This change modifies things to take a different approach. Instead of
using run_deployment, we will now standup an SQS handler that will
invoke the functions directly. Then, instead of calling run_deployment
directly, we will send an SQS to the handler. This way, we don't have to
deal with the overhead of going back and forth from the prefect server
and worker just to start a task. This also has the additional benefit of
re-using the same process for each task that we execute. This is a plus
in terms of latency because "init" code can run once at application
startup instead of once per flow.

* PR feedback