[WC-126] Kick off regression test on pushes to main and build-me commits (#225)
This change adds a script, run_regression.sh, that will get called
whenever a commit is merged to main or a commit contains the "build-me"
string. It can also be invoked locally. It will send a message to
boosted dag to trigger a run of the regression tests.

Additionally, this change adds a new field to the regression output,
execution_plan_simple, that will make it easier to diff runs. We also
re-enable the argument and output logging that was added earlier. This
was removed because it was spammy during local runs. It is disabled
during local runs now, so should not be a problem.