[WC-322] Log sqs messages to clickhouse, sqs exception handling (#244)
Add log_event calls to the sqs handler main. This will allow us to
better audit messages that come into our system. Additionally, add a
generic exception handler to the sqs handler. Not only will this allow
us to see exceptions in clickhouse, but this will make the handler not
need to restart whenever there is an exception.