[WC-322] Add send_time_utc to SQS messages (#246)
This will allow us to detect when the sqs handler is backed up. The
difference between the start_time_utc and send_time_utc in
agent.worker_sqs_log will be equal to the "wait time", that is how long
the message was sitting in the SQS queue before we picked it up.
Ideally, this difference is as small as possible.