Updated cancel logic - manage our own state (#289)
* WIP Updated cancel logic - manage our own state

This change implements a custom cancellation logic. Instead of relying
on prefect, we will keep track of which jobs are cancelled in the table
agent.cancelled_ids. These ids are either an execution_plan_id or a
plan_run_id, this is why the table is named generically. Then, jobs will
be responsible for checking if they are cancelled and aborting if
necessary.

It also updates the logic to display the "Cancelled" status on the
frontend correctly.

* small fixes

* lint

* comment

---------

Co-authored-by: Zach <zach.silver@gradientboostedinvestments.com>