[SR24-1106] Use a single GPTService connection for all requests (#387)
Creating a new GPTService connection on each requests creates a lot of
overhead, especially when we are making tens of queries in parallel.
We've gone back and forth on using a single global connection or a
new connection for each request. A singal global connection performs
much better than a new connection for each request, but was not working
properly when we were using prefect "deployments". Now that we are not
using these, we can switch back to using a single global connection,
which should improve stability a lot. I'm enabling this by default, but
this can be reverted by setting any non empty value to the env var
NEW_GPT_SERVICE_CONNECTION_FOR_EACH_REQUEST.