[SR24-1106] Disable USE_GLOBAL_STUB by default (#389)
We just enabled this by default, but it is causing issues with the agent
web server since the event loop handling is a little sloppy. Let's
disable this by default and make the env var that controls it the same
as the name of the global. We'll explicitly enable this for the workers.