[SR24-1106] Better GPTService connection sharing (#409)
We've been trying to have a global GPTService connection for all of
agent service. The reason we have been trying to do this is because
there is a lot of overhead associated with creating a new connection to
the service, and this is especially bad when we are making hundreds of
requests in parallel. The global stub approach was causing serious
issues with prefect - something in the prefect runner is killing the
active event loop and making the connection unusable.

We could refactor a ton of code to make using a global stub possible in
prefect, but in the meantime we can scope the connection to the GPT
instance. This should get us a lot of the benefit of the global stub but
should also work with prefect, since we'll be creating a new connection
for each prefect task.