[SR24-1106] Add event logging for new gpt service connections (#413)
Despite modifying the GPTService connection logic, we are still running
into a lot of issues where clients are unable to connect to GPT service.
It seems like we are still creating a lot of connection in loops, which
is causing a lot of issues. Let's add some event logging so that we have
visibility into this.