[WC-685] Share GPTService connections for hypothesis tasks (#426)
We were able to see that some of the hypothesis code was creating a ton
of GPTService connections by looking at
agent.gpt_service_connection_created in Clickhouse. Specifically, this
was a big problem in test_hypothesis_for_news_developments. This change
modifies the Hypothesis code to optionally take a GPTServiceStub. This
way, we can share the connection for all the relevant queries which will
improve stability. There are a few other places where this code is
called, so let's update it everywhere.