[WC-558] Optionally run run_execution_plan tasks with boosted dag (#447)
There is a problem currently where jobs being processed by the sqs
handler get terminated whenever we deploy a new version of agent
service. We have a 30 minute grace period that helps a little bit with
this, but we have consistenly been seeing issues where jobs get killed
during a deploy. When this happens, there is new automated retry logic.
Additionally, if all of the sqs handler instances are backed up
processing long running run_execution_plan jobs it can slow down the
entire system since we do not have a way to dynamically provision more
hosts based off of the size of an sqs queue.

As a potential solution to these issues, we'd like to try running each
task as its own pod in k8s by using Boosted DAG/Argo. Instead of running
an sqs handler that processes all of the messages, we can submit each
sqs message as its own job in argo. While this can introduce some
overhead to run_execution_plan calls, this should be ok for two reasons:

1. run_execution_plan jobs can run very long, so the overhead is not as
important as for create_execution_plan
2. It should be possible to make this overhead minimal by keeping a pool
of "warm" hosts up that always contain the latest code image.

This change adds this behavior behind a feature flag,
boosted-dag-run-execution-plan. When this flag is "true", we will send
the `run_execution_plan` sqs messages to the bootsed dag queue instead
of to the agent worker queue. This change also adds the main
`sqs_execute.py`. This will be invoked by argo. Note, because the
run_execution_plan message can be quite large (approaching megabytes in
size) and because workflow yamls have a strict size limit,
sqs_execute.py can optionally take an s3 path to download the message
from.