[WC-319] Add extra debug info to tool calls (#482)
Currently whenever a tool is called we log all of its arguments and the
result of the tool call to clickhouse as json. This is very helpful for
debugging, but there are cases where additional debug info can also be
extremely helpful. Some examples are:

- chatGPT output (ie, code it generates for transform_table)
- other service response (ie, content of PA service response)

This change makes it possible to add extra debug info using context
vars. The context var TOOL_DEBUG_INFO can be imported and modified
during a tool call. After the call completes. the contents of this
variable will get added to a new column in agent.tool_calls.

Unrelated: another fix for sqs logging