Update versions to fix psycopg pool conflict (#539)
psycopg_pool.AsyncConnectionPool.check_connection exists in psycopg-pool
versions 3.2 and up. Recently, a change was made to gbi-common-py-utils
to require ~3.1 in its setup.py. This caused the app to be unusable,
since connections were not able to be created at all to the DB.

This change modifies the versions of gbi-common-py-utils and
boosted-data-access-layer to make sure we do NOT get the 3.1 requirement
in the setup.py of gbi-common-py-utils. Note that the dal downgrade is
necessary since its latest versions requires a version of
gbi-common-py-utils that has the 3.1 requirement.

Also, add an exeplicit dependency on psycopg-pool verison ~3.2.

Also, add the stacktrace to the `ValueError` when we can't create a
connection.