[WC-1004] 'agent-canned-prompt' clickhouse event (#618)
On the main agent screen, there is a list of a few "canned" prompts that
the user can click on as examples. These are heavily used for new
clients and during onboarding, and we need to make sure that they always
work properly. Currently, it is a bit difficult to determine which
agent_ids come from canned prompts, which in turn makes it difficult to
audit our performance on these.

This change adds a new event, agent-canned-prompt, that will get logged
whenever a ChatWithAgentRequest contains the new "canned_prompt_id"
field.

After we start collecting this, we can modify some of the clickhouse
audit tables so that they get this data.