[WC-1105] Increase concurrent filter calls for hypothesis tools (#674)
filter_stocks_by_product_or_service and
analyze_hypothesis_with_categories are often times a bit slow. Both of
these tools are commonly used in "hypothesis" type queries which are
very important for our users. We are trying to make them as fast as
possible.

Previously, GPTService could only handle 100 concurrent requests on a
single connection. Now, this has been increased to 2000. This means we
can safely increase the concurrency of some of the tools and should
hopefully see some performance increases. Let's start this at 300, and
then if things look good we can further increase this.