[WC-1178] Move canned prompts to agent server (#717)
We want to make it easier to ensure that canned prompts always have
regression tests. This change adds a new endpoint, /canned-prompts, that
will return the list of "canned" prompts that get displayed on the
homescreen for users to test with. It also modifies
TestHomeScreenQueries to use these canned prompts more explicitly.

Still need to find a good way to enforce that canned prompts always have
regression tests. This is at least a good start, because we can be
confident that if a canned prompt is modified it will get reflected in
the regression test. But we still need a way to enforce that a
regression test exists for all canned prompts.