[WC-1221] Implement soft delete, restore agent (#725)
Currently when an agent is deleted remove it from the agents table in
Postgres. This nukes the agent and all of its data, which makes it
impossible to revisit or restore the agent. This change modifies the
delete logic to instead set a flag in the agents table marking the
deleted agent as deleted. This will make the agent invisible to users,
but the agent will still be accessible via the link.

Additionaly, the change adds a new post endpoint, /restore-agent, that
will "undo" the delete.