[QL24-15] Add "real_user_id" to User, include in Message (#728)
We want to be able to track when a request is made by a spoofed user.
The frontend will start sending a "realuserid" header that will include
the user_id of the caller, that might differ from the user id included
in the JWT. This PR takes this field and puts it into the User object.
This will be logged in clickhouse for debugging.

We also want to include the real_user_id as the "message_author" in the
Message object. This will allow us to detect when messages were sent by
CS on behalf of a client.