[WC-1460] Use gpt-4o-mini as first pass in filter_stocks_by_profile_match (#802)
filter_stocks_by_profile_match is the most expensive tool currently.
This is due to the fact that it uses Claude Sonnet to filter a list of
stocks and generate explanations which ends up being expensive due to
the number of stocks that it needs to filter.

gpt-4o-mini is an order of magnitude cheaper than Sonnet, but is not
smart enough to write good explanations and handle the citation logic.
However, it is still very good at simple filtering tasks. This change
adds a "first pass" round of filtering to this tool that uses
gpt-4o-mini. Then, we send the results of this initial filter to Sonnet
for further processing. This should result in a significant decrease in
the number of tokens sent to Sonnet. This should hopefully reduce
latency and costs with this tool.