[WC-1480] #alfa-client-queries slack updates (#817)
We are continuing to monitor Alfa's performance on client queries. As
part of this, memeber of CS and Product would like to be able to get
updates whenever clients are creating new agents. Specifically, they'd
like to get a slack message with all of the details of the agent sent to
a channel six hours after the agent is first created.

This change adds the code necessary to send the slack message. Instead
of importing any slack libraries, I opted to use the requests library
directly. I also added event logging so we'll be able to see if there
are any issues from the slack API.

Note that we only want to send the message for client queries. This
means that if the user that created the agent is a boosted user we do
NOT want to send the message.