[WC-1545] Copy analyst (#904)
Add endpoint for analyst copy. This is pretty similar to how model copy
works from Boosted 1.

100% of the data from the original analyst will be copied. At the moment
the copy completes successfully, the two agents will be identical.
However, the following ids will all be different:

- agent_id
- plan_id(s)
- plan_run_id(s)
- output_id(s)
- work_log_entry_id(s)

Any new messages sent to either the original or the newly copied agent
will only affect the agent they are sent to - the agents are completely
independent.

Note: this is only avilable for admins. At some point, we will likely
add logic to "block" copying agents that use "prohibited" tools. An
example might be the custom docs tool - we don't want to accidentally
expose custom docs of a user.