[QL24-1855] Update agent output caching (#1148)
- Cache each individual output_id instead of the GetAgentOutputResponse.
  The GetAgentOutputResponse contains data that is mutable, and we were
	- Note there is a serious bug in pydantic that causes a lot of
	  outputs to not serialize properly. There is apparently a fix
incoming for this - we will upgrade the pydantic version ASAP when this
happens. For outputs where this happens, they will still work properly,
but they will not get cached.
running into inconsistency issues because of this.
- Make cache usage optional. If the REDIS_HOST env var is set and the
  feature flag agent-output-cache is set to true, then the cache will be
enabled. Note, the service will need to be restarted for this to take
effect.