Fix regression  tests (#1189)
For some reason, after we modified the decorators used for
run_execution_plan, the execution_log argument was no longer getting
returned properly to the original caller. This argument was modified in
place and then later used by the regression test framework to confirm
that particular tools were ran.

The fix for this is to explicitly return the modified execution_log. I'm
honestly not sure why this works, but it fixes the issue.