[WC-2096] Error on "empty output" for run_execution_plan (#1233)
In addition to logging to slack whenever a live plan fails, we also want
to log to slack when a live plan produces an "empty" output. "empty"
means different things for different IOTypes.

While we are at it - we should return a nonzero exit code if the output
that is produced is empty.