[WC-2252] Inline GPTService (#1379)
This changes moves all of the logic from the GPTService GRPC Service
directly into agent service via the llm-client library. The motivations for this are twofold:

1. We have been seeing more and more stability issues related to the
number of queries that GPTService is getting from clients all across our
system. While scaling up can help this, we think we can improve latency
and stability by removing this middleman entirely

2. OpenAI and Anthropic continue to release updates to their API. One in
parricular that OpenAI recently added, structured outputs, is extremely
useful. However, it is not well suited for a GRPC RPC. See the docs
here https://platform.openai.com/docs/guides/structured-outputs

This change inlines the llm client code and also adds explicit support
for the OpenAI structured outputs features. ALL of the features from the
GPTService GRPC service are included in this change. This includes:

- caching (even for structured outputs)
- rate limiting
- audit logging