Increase number of retries on LLM client (#1427)
Recently (starting november 1st 2024) we noticed a big uptick in openAI
connection refused errors with gpt-4o-mini. These errors are related to
the TEXT_SNIPPET_RELEVANCY_MAIN_PROMPT prompt which consumes billions of
gpt-4o-mini tokens per day. This is not too surprising - we get 30,000
requests per minute and 150,000,000 tokens per minute so we are going to
be bottlenecked by this to some extend.

When we removed GPTService from this repo, we also reduced the total
number of retries on LLM calls since we removed the GRPC retry
decorator. We should add this back since we are seeing more errors
surface during agent runs.