[WC-2490] Add progress bar option to gather_with_concurrency (#1712)
* [WC-2490] Add progress bar option to gather_with_concurrency

We have a lot of tools that iterate through a list of stocks, texts etc
and can take a long time to finish. Often it becomes difficult to tell
if these agents are "stuck" or not by just looking at the stdout logs
because they do not always output status updates. Ideally, we'd have
some progress bar that could make this much easier visually to see.

Luckily, there is a perfect library for this - tqdm. This changes
modifies the gather_with_concurrency function by adding two new optional
params - use_progress_bar and desc. When use_progress_bar is set to
true, the code will be wired into tqdm which will output a nice progress
bar. The desc param controls the text associated with the progress bar.

There are tons of places where we call into gather_with_concurrency, so
this change makes the progress bar opt-in and only applies it directly
to a few places.

* Update agent_service/tools/stock_filter_and_rank/utils.py

Co-authored-by: David Grohmann <104018429+davidgrohmann@users.noreply.github.com>

---------

Co-authored-by: David Grohmann <104018429+davidgrohmann@users.noreply.github.com>