[WC-2490] Add progress bar logic for frontend (#1753)
In a previous change we added the tqdm library to the
`gather_with_concurrency` function. This allowed us to display a
progress bar in stdout for various functions. This is helpful, but now
we'd like to export this progress bar to the frontend. In order to do
this, we hook into the tqdm library and add custom logic that gets run
whenever the bar gets updated. We send out agent events with the
progress that the frontend can use to display a progress bar.